refactor(ds): extract MatterInspector to features/matters (5d.4b)
Third of the per-feature inspector splits. Moves the matter render
block (~380 lines) + all matter editor state + matter helpers out of
InspectorPanel into a focused feature-local component.

MatterInspector (717 lines)
---------------------------
- src/features/matters/components/MatterInspector.tsx
- Consumes useMatterDetail (5d.1)
- Consumes identity helpers from shared inspector/identityHelpers (5d.4a)
- Owns matter editor state:
  - activeMatterEditor (12-position discriminator)
  - isSavingMatterStatus, isSavingMatterField
  - inspectorMatterStatus + setInspectorMatterStatus
- Owns the 14 resolvedMatter* field computations + the assignee resolvers
  + matterTeamIdentities memo
- Owns handleMatterStatusChange + handleMatterPatchChange (with the
  camelCase → snake_case keyMap for the patch normalization)
- Owns matterStatusOptions + urgencyOptions memos
- Renders its own loading skeleton

InspectorPanel dispatcher update
--------------------------------
- Imports + delegates: <MatterInspector practiceId={...} entityId={...}
  {...all matter props} /> when entityType === 'matter' and both ids
  are present
- Removed 320+ lines of matter-specific state/memos/helpers/handlers:
  - State: activeMatterEditor, isSavingMatterStatus, isSavingMatterField,
    inspectorMatterStatus + sync effect
  - Memos: matterStatusOptions, matterClientOptionsWithNone,
    matterTeamIdentities, matterUrgencyLabel, urgencyOptions,
    resolvedMatterClientLabel, resolvedMatterAssigneeIds,
    resolvedMatterAssigneeNames
  - Resolved fields: clientName, clientId, urgency, responsibleAttorneyId,
    originatingAttorneyId, caseNumber, type, court, judge, opposingParty,
    opposingCounsel, createdLabel, updatedLabel, billingLabel
  - Helpers: resolveMatterClientIdentity, resolveAttorneyLabel,
    resolveAttorneyIdentity (local wrapper — helper itself stays in
    identityHelpers for ConversationInspector to use in 5d.5)
  - Handlers: handleMatterStatusChange, handleMatterPatchChange
- Removed canEditMatterStatus / canEditMatterFields gates (consumed
  only by removed render block)
- Removed setActiveMatterEditor call from the reset useEffect
- Removed loading skeleton block for matter (MatterInspector renders
  its own)
- Dropped unused imports: InspectorHeaderEntity, MatterFilesSection,
  MATTER_STATUS_LABELS, MATTER_WORKFLOW_STATUSES, isValidMatterStatus
  helper, resolveAttorneyLabelHelper, resolveAttorneyIdentityHelper,
  useCallback
- matterBillingLabel prop kept on InspectorPanelProps for caller
  compatibility but underscored in destructuring (unused now)

Net InspectorPanel change: -628 lines deleted, +89 lines added (the
MatterInspector delegation + comments)

Conversation block (still in InspectorPanel) retains its hold on:
- inspectorMatterStatus and setInspectorMatterStatus → removed since
  the conversation render didn't reference them (only the matter block
  did)
- isSavingMatter (separate from isSavingMatterStatus — used by the
  conversation's matter-combobox editor)
- resolveAttorneyIdentity → call sites in conversation block stayed
  via the helper import, not the local wrapper
